# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"userId": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"userId"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "REGISTER USER",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"token": {}
													},
													"required": [
														"token"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "LOGIN",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "FORGOT PASSWORD (Send reset token via email)",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "RESET PASSWORD",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newPassword": {
										"type": "string"
									},
									"token": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"newPassword",
									"token",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/profile": {
			"post": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"email": {
															"type": "string"
														},
														"username": {
															"type": "string"
														}
													},
													"required": [
														"email",
														"username"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "UPDATE PROFILE",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newPassword": {
										"type": "string"
									},
									"currentPassword": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/admin/users": {
			"post": {
				"operationId": "CreateAdminUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"userId": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"userId"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "CREATE ADMIN USER (internal use)",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAdminUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"users": {
													"items": {
														"properties": {
															"updatedAt": {
																"type": "string",
																"format": "date-time"
															},
															"username": {
																"type": "string"
															},
															"email": {
																"type": "string"
															},
															"createdAt": {
																"type": "string",
																"format": "date-time"
															},
															"id": {
																"type": "number",
																"format": "double"
															}
														},
														"required": [
															"updatedAt",
															"username",
															"email",
															"createdAt",
															"id"
														],
														"type": "object"
													},
													"type": "array"
												}
											},
											"required": [
												"users"
											],
											"type": "object"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "GET ALL USERS",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/admin/users/{userId}": {
			"delete": {
				"operationId": "DeleteAdminUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "DELETE USER (Prevent self-deletion)",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}